🚚 Move : comments 관련 컴포넌트들 productdetail 폴더로 이동 #53
